raspi-home/requirements.txt
Flask
requests
opencv-python
hx711
RPi.GPIO
